requirements.txt
alembic==1.5.8
aniso8601==8.1.0
applicationinsights==0.11.9
asgiref==3.2.10
azure-core==1.17.0
azure-functions==1.7.2
azure-iot-device==2.5.1
azure-iot-hub==2.5.0
azure-storage-blob==12.8.1
bcrypt==3.2.0
beautifulsoup4==4.5.3
bson==0.5.8
cached-property==1.5.2
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.4.6
cycler==0.10.0
decorator==4.4.2
deprecation==2.1.0
distro==1.6.0
Django==3.1
dnspython==2.1.0
docker==4.4.1
docker-compose==1.28.3
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docx2pdf==0.1.7
email-validator==1.1.2
face-recognition-models==0.3.0
Flask==1.0.2
Flask-API.yandex==0.6.2.1
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
httpie==1.0.0
idna==2.7
image==1.5.32
imageio==2.9.0
img2pdf==0.4.0
importlib-metadata==1.7.0
iotedgehubdev==0.14.9
isodate==0.6.0
itsdangerous==1.1.0
janus==0.4.0
Jinja2==2.11.2
joblib==0.16.0
jsonpath-rw==1.4.0
jsonschema==2.6.0
kiwisolver==1.2.0
Lasagne==0.1
lxml==4.6.3
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.3.0
msrest==0.6.21
networkx==2.4
nltk==3.5
numpy==1.19.1
oauthlib==3.1.1
opencv-python==4.3.0.38
packaging==20.9
paho-mqtt==1.5.1
paramiko==2.7.2
pdf2image==1.13.1
pdf2jpg==1.0
pdfkit==0.6.1
pikepdf==1.18.0
Pillow==7.2.0
ply==3.11
pycparser==2.20
Pygments==2.7.2
pymongo==3.7.1
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pypiwin32==223
PySocks==1.7.1
pytesseract==0.3.5
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.19.0
python-editor==1.0.4
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
PyYAML==4.2b4
regex==2020.7.14
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
scikit-image==0.17.2
scipy==1.5.2
six==1.15.0
SQLAlchemy==1.3.23
sqlparse==0.3.1
texttable==0.9.1
Theano==1.0.5
tifffile==2020.7.24
tqdm==4.48.2
uamqp==1.4.1
urllib3==1.24.3
websocket-client==0.57.0
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.0
